Supervisor: - Remove Mongo configuration - Add Influx configuration Assembly: - Send one sensor data message at application startup to automatically create a queue in AMQP with Spring Boot Systems: - Adapt docker-compose.prod.yaml to create influx container - Make systems wait for RabbitMQ (AMQP) startup Experiment: - Create Python script for automatically conducting the experiment

+++ src/systems/assembly/src/main/java/dk/sdu/se23g6/arch/assembly/production/SensorDataSender.java
@@ -57,7 +57,7 @@ public class SensorDataSender {
 
     private String calculateExpectedMessages() {
         long millis = Duration.between(experimentStart, experimentEnd).toMillis();
-        double seconds = millis / 1000d;
+        long seconds = millis / 1000;
         return "Duration: " + DurationFormatUtils.formatDurationHMS(millis)
                 + "\nExpected Entries: " + this.messagesPerSecond * (seconds + 1) + "\n";
     }
